Add command and arguments options; remove forget
This commit adds the -c and -p options, which specify the backup
command to run and the arguments to pass to it. It also removes the -N
option for running restic in dry run mode because you can now
accomplish the same thing with -p "-n". If -c is not specified, the
command `/usr/local/bin/restic backup` is run. If -p is not specified,
no additional arguments are passed to the command, just the path.

Finally, this commit also removes the `restic forget` command from the
script. Because it doesn't require taking or mounting zfs snapshots, it
should be run separately. For example:
    restomic /usr/home && restic forget --keep-last 7 --prune